holly shit
we got a three

+++ Structures.h
@@ -5,8 +5,8 @@
 
 
 struct position {
-    int x{};
-    int y{};
+    float x{};
+    float y{};
 };
 
 struct IParam {
